guizero/ultraschall_sensor/gui/requirements.txt
pyserial==latest
guizero==latest